// server/types/db-tables.ts
export enum DbTables {
  schedule = "schedule",
  payment = "payment",
  student = "student",
  event = "event",
  event_category = "event_category",
  color = "color",
  english_level = "english_level",
  purpose_lesson = "purpose_lesson",
  contact = "contact",
  order_platform = "order_platform",
  users = "users",
  settings = "settings"
}